requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.26.125
botocore==1.29.125
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.29.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.12
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.3
